requirements.txt
flask
Flask-SQLAlchemy
python-dotenv
Flask-HTTPAuth
flask-marshmallow
flask-restful
flask-bootstrap
psycopg2
